+++ lbc/problems/example2.py
@@ -5,8 +5,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from lbc import plotter
-from lbc.problems.problem import Problem
-from lbc.util import contains
+from lbc.problems.problem import Problem, contains
 
 
 class Example2(Problem):
