+++ lbc/problems/example9.py
@@ -6,8 +6,7 @@ import matplotlib.patches as patches
 import matplotlib.pyplot as plt
 
 from lbc import plotter
-from lbc.problems.problem import Problem
-from lbc.util import sample_vector, contains
+from lbc.problems.problem import Problem, contains, sample_vector
 
 
 class Example9(Problem):
