+++ lbc/problems/problem.py
@@ -1,4 +1,15 @@
-from lbc.util import sample_vector
+import numpy as np
+
+def contains(vector, lims):
+    return (vector[:, 0] >= lims[:, 0]).all() and (vector[:, 0] <= lims[:, 1]).all()
+
+def sample_vector(lims, damp=0.0):
+    # from cube
+    dim = lims.shape[0]
+    x = np.zeros((dim, 1))
+    for i in range(dim):
+        x[i] = lims[i, 0] + np.random.uniform(damp, 1 - damp) * (lims[i, 1] - lims[i, 0])
+    return x
 
 
 class Problem:
