+++ lbc/scripts/run.py
@@ -6,8 +6,7 @@ from queue import Queue
 from lbc import plotter
 from lbc.learning.oracles import get_oracles
 from lbc.param import Param
-from lbc.problems.utils import get_solver, get_problem
-from lbc.util import init_tqdm, update_tqdm
+from lbc.util import init_tqdm, update_tqdm, get_problem, get_solver
 
 
 def make_instance(param):
